Change _tensor_schema module to package
